deploy/requirements.txt
# Requirements for SageMaker job submission scripts
boto3>=1.34.0
pyyaml>=6.0
